+++ cmd/messaging/domain/movie/model/movie.go
@@ -1,0 +1,1 @@
+package model
