+++ cmd/messaging/domain/movie/service/movie_service.go
@@ -1,0 +1,1 @@
+package service
