+++ cmd/messaging/line_messaging_service.go
@@ -1,0 +1,9 @@
+package messaging
+
+func PublishMessage() {
+
+	// メッセージを作成する
+
+	// メッセージを送信する
+
+}
